+++ 02-Trunk/grosup-practice-war/src/main/resources/applicationContext.xml
@@ -33,15 +33,18 @@
          <property name="username" value="${jdbc.username}" />
          <property name="password" value="${jdbc.password}" />
 		<!-- 初始化连接大小 -->
-         <property name="initialSize" value="${jdbc.initialSize}" />
+<!--          <property name="initialSize" value="${jdbc.initialSize}" /> -->
          <!-- 连接池最大数量 -->
          <property name="maxActive" value="${jdbc.maxActive}" />
          <!-- 连接池最大空闲 -->
          <property name="maxIdle" value="${jdbc.maxIdle}" />
-         <!-- 连接池最小空闲 -->
-         <property name="minIdle" value="${jdbc.minIdle}" />
+<!--          连接池最小空闲 -->
+<!--          <property name="minIdle" value="${jdbc.minIdle}" /> -->
          <!-- 获取连接最大等待时间 -->
          <property name="maxWait" value="${jdbc.maxWait}" />
+         <property name="defaultAutoCommit" value="true"></property>
+		<property name="testWhileIdle" value="true"></property>
+		<property name="validationQuery" value="select 1 from dual"></property>
 	</bean>
 	
 	<bean id="sqlSessionFactory" class="org.mybatis.spring.SqlSessionFactoryBean">
