+++ 02-Trunk/grosup-practice-war/src/main/resources/jdbc1.properties
@@ -2,8 +2,6 @@ jdbc.driver=com.mysql.jdbc.Driver
 jdbc.url=jdbc:mysql://120.27.46.102:3306/practice?useUnicode=true&amp;characterEncoding=UTF-8&amp;zeroDateTimeBehavior=convertToNull
 jdbc.username=root
 jdbc.password=practice
-jdbc.maxActive=60
-jdbc.maxIdle=30
+jdbc.maxActive=30
+jdbc.maxIdle=10
 jdbc.maxWait=500
-jdbc.initialSize=20
-jdbc.minIdle=20
