--- 02-Trunk/grosup-practice-war/src/main/resources/mybatis/test-orm.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.practice.test">
-
-	<resultMap type="com.grosup.practice.beans.UserBean" id="BaseResultMap">
-		<result column="NAME" jdbcType="VARCHAR" property="name"/>
-		<result column="AGE" jdbcType="INTEGER" property="age"/>
-	</resultMap>
-	
-	<select id="queryTest" resultMap="BaseResultMap">
-		select * from user where id = #{id}
-	</select>
-</mapper>
