+++ 02-Trunk/grosup-practice-war/src/main/resources/sqlMapConfig.xml
@@ -4,7 +4,6 @@
 <configuration>
 	<mappers>
 		<!-- <mapper resource="mybatis/xxxx-orm.xml" /> -->
-		<mapper resource="mybatis/test-orm.xml"/>
 		<mapper resource="mybatis/student-orm.xml"/>
 		<mapper resource="mybatis/class-orm.xml"/>
 		<mapper resource="mybatis/user-orm.xml"/>
